Changed grid size
By making the grid size a power of 2 times 3 times the number of SMs, there is less branching during integer exponentiation (as the last few digits of the exponent are identical in binary for threads in the same block). The exact power of 2 may vary based on the size of the input. However, experimentally, 16 was the most efficient.