Switched to fixed-point mathematics
Instead of utilising doubles to represent the fractional part of our  partial sum, we multiply it by 2^64 and write it in an unsigned 64-bit integer instead. This simplifies the code by removing the need to reduce modulo 1 every addition step as this is naturally handled by integer overflow, and allows for 11 more bits of precision.